fix(ui): band long durations in the message timing footer
A two-hour turn rendered "8158.94s" in the footer and its tooltip: the
footer had its own formatter that divided by 1000 and stopped there, while
the thread's running indicator banded the same duration into "2h 15m".

Move both readouts into features/chat/format-duration.ts. Sub-minute turns
keep their fractional second — the difference between a 2s and a 9s turn is
worth reading — and past a minute the footer hands off to the same banded
formatter the live indicator uses.
