+++ .changeset/loud-pans-shave.md
@@ -1,0 +1,5 @@
+---
+'@qlan-ro/mainframe-ui': patch
+---
+
+Band long turn durations in the message timing footer. A two-hour turn read "8158.94s"; it now reads "2h 15m", matching the running indicator's own readout. Turns under a minute keep their fractional second.
