+++ packages/ui/src/features/chat/thread/ChatThread.tsx
@@ -22,7 +22,8 @@ import { CompactingPill } from '../messages/SystemMessage';
 import { DegradedChatCard } from './DegradedChatCard';
 import { useChatExtras } from '../runtime/use-chat-thread-runtime';
 import { useRotatingPhrase } from './use-rotating-phrase';
-import { formatElapsedSeconds, useRunElapsed } from './use-run-elapsed';
+import { formatElapsedSeconds } from '../format-duration';
+import { useRunElapsed } from './use-run-elapsed';
 import { SkillsProvider } from '@/features/skills/use-chat-skills';
 import { FindBar } from '../find/FindBar';
 import { useFindHotkey } from '../find/use-find-hotkey';
