+++ packages/ui/src/features/chat/thread/__tests__/use-run-elapsed.test.ts
@@ -6,41 +6,15 @@
  * formatter's own arithmetic.
  *
  * Behaviors covered:
- *  - the formatter's three bands (seconds, minutes, hours) and their boundaries
  *  - nothing is reported below 1s (a sub-second run must not flash "0s")
  *  - the count ticks once per second while a run is active
  *  - it clears when the run ends, and a second run restarts from zero
+ *
+ * The readout's formatting lives in `features/chat/__tests__/format-duration`.
  */
 import { describe, expect, it, beforeEach, afterEach, vi } from 'vitest';
 import { act, renderHook } from '@testing-library/react';
-import { formatElapsedSeconds, useRunElapsed } from '../use-run-elapsed';
-
-describe('formatElapsedSeconds', () => {
-  it('reports whole seconds below a minute', () => {
-    expect(formatElapsedSeconds(0)).toBe('0s');
-    expect(formatElapsedSeconds(1)).toBe('1s');
-    expect(formatElapsedSeconds(9)).toBe('9s');
-    expect(formatElapsedSeconds(59)).toBe('59s');
-  });
-
-  it('switches to minutes at 60s, zero-padding the seconds so a ticking readout keeps its width', () => {
-    expect(formatElapsedSeconds(60)).toBe('1m 00s');
-    expect(formatElapsedSeconds(65)).toBe('1m 05s');
-    expect(formatElapsedSeconds(75)).toBe('1m 15s');
-    expect(formatElapsedSeconds(600)).toBe('10m 00s');
-    expect(formatElapsedSeconds(3599)).toBe('59m 59s');
-  });
-
-  it('drops seconds past an hour', () => {
-    expect(formatElapsedSeconds(3600)).toBe('1h 00m');
-    expect(formatElapsedSeconds(3725)).toBe('1h 02m');
-    expect(formatElapsedSeconds(7380)).toBe('2h 03m');
-  });
-
-  it('clamps a negative reading to zero', () => {
-    expect(formatElapsedSeconds(-5)).toBe('0s');
-  });
-});
+import { useRunElapsed } from '../use-run-elapsed';
 
 describe('useRunElapsed', () => {
   beforeEach(() => {
